chapter_4_production_polish/backend/pyproject.toml
[project]
name = "voice-escape-room-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "python-dotenv",
]
